CReating Scene01
For achor based AR i have creaed a whole new scene from the scene00

+++ AR00/Assets/Scripts/GPS.cs
@@ -10,8 +10,8 @@ public class GPS : MonoBehaviour
     public static GPS Instance { set; get; }
 
 
-    public float latitude;
-    public float longitude;
+    public double latitude;
+    public double longitude;
 
 
     private void Start()
